requirements.txt
django==4.2.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-split-settings==1.2.0
